--- src/jit_no_ir.cpp
@@ -1,1 +1,0 @@
-#include "jit_no_ir.h"
